refactor: WaitingWithRank 클래스 위치 변경 및 패키지 구조 정리
- WaitingWithRank를 domain 패키지에서 infrastructure.dto 패키지로 이동
- 관련된 클래스 및 테스트 파일의 import 경로 수정
- 불필요한 import 제거로 코드 정리
