+++ src/main/java/roomescape/reservation/infrastructure/JpaWaitingRepository.java
@@ -3,7 +3,7 @@ package roomescape.reservation.infrastructure;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
-import roomescape.reservation.domain.WaitingWithRank;
+import roomescape.reservation.infrastructure.dto.WaitingWithRank;
 import roomescape.reservation.domain.Waiting;
 
 import java.util.List;
@@ -13,7 +13,7 @@ public interface JpaWaitingRepository extends CrudRepository<Waiting, Long> {
     boolean existsByReservationIdAndMemberId(Long reservationId, Long memberId);
 
     @Query("""
-        SELECT new roomescape.reservation.domain.WaitingWithRank(
+        SELECT new roomescape.reservation.infrastructure.dto.WaitingWithRank(
             w,
             (SELECT COUNT(w2) + 1
              FROM Waiting w2
